Excel Automation
Create a process using the Workbook and DataTable activities to do the following tasks: 

Add a new column to an existing Excel file that'll contain the estimated value of the product. 

Calculate the estimated value by multiplying the Quantity with the Price per Kg. 

Update the workbook with the newly updated data in a new sheet.